too a long rest, am now back to usual silly 3D engine business

+++ src/engine_3D/rotations.c
@@ -13,7 +13,7 @@
  * @param vect 		2D vector
  * @param angle_rad angle in radians
  */
-void	rotate_vector(double *v, double angle_rad)
+void	rotate_vector(float *v, float angle_rad)
 {
 	v[X] = v[X] * cos(angle_rad) - v[Y] * sin(angle_rad);
 	v[Y] = v[X] * sin(angle_rad) + v[Y] * cos(angle_rad);
